task3/code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
appdirs=1.4.4=pypi_0
audioread=2.1.9=pypi_0
blas=1.0=mkl
ca-certificates=2020.10.14=0
certifi=2020.12.5=py36haa95532_0
cffi=1.14.4=pypi_0
chardet=4.0.0=pypi_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=py36h009560c_0
decorator=4.4.2=pypi_0
docutils=0.16=pypi_0
freetype=2.10.4=hd328e21_0
fsspec=0.8.4=pypi_0
icu=58.2=ha925a31_3
idna=2.10=pypi_0
intel-openmp=2020.2=254
joblib=0.17.0=pypi_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.0=py36hd77b12b_0
libpng=1.6.37=h2a8f88b_0
librosa=0.8.0=pypi_0
libtiff=4.1.0=h56a325e_1
llvmlite=0.35.0=pypi_0
lz4-c=1.9.2=hf4a77e7_3
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py36hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py36hb782905_0
mkl_fft=1.2.0=py36h45dec08_0
mkl_random=1.1.1=py36h47e9c7a_0
munkres=1.1.4=pypi_0
ninja=1.10.2=py36h6d14046_0
numba=0.52.0=pypi_0
numpy=1.19.2=py36hadc3359_0
numpy-base=1.19.2=py36ha3acd2a_0
olefile=0.46=py36_0
opencv-contrib-python=4.4.0.46=pypi_0
openssl=1.1.1h=he774522_0
packaging=20.8=pypi_0
pandas=1.1.4=pypi_0
pillow=8.0.1=py36h4fa10fc_0
pip=20.2.4=py36haa95532_0
pooch=1.3.0=pypi_0
pycparser=2.20=pypi_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py36h6538335_2
python=3.6.12=h5500b2f_2
python-dateutil=2.8.1=py_0
pytorch=1.6.0=py3.6_cuda101_cudnn7_0
pytz=2020.4=pypi_0
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pypi_0
resampy=0.2.2=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.4=pypi_0
setuptools=50.3.0=py36haa95532_1
sip=4.19.8=py36h6538335_0
six=1.15.0=py_0
soundfile=0.10.3.post1=pypi_0
sqlite=3.33.0=h2a8f88b_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=he774522_0
torchvision=0.7.0=py36_cu101
tornado=6.0.4=py36he774522_1
tqdm=4.55.0=pypi_0
urllib3=1.26.2=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wheel=0.35.1=py_0
wincertstore=0.2=py36h7fe50ca_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
